Fix compilation error: Number cannot be converted to Long
Add .longValue() call to convert Number to Long in XssSecurityTest.java line 304.

+++ backend/src/test/java/nl/ticketservice/resource/XssSecurityTest.java
@@ -335,7 +335,7 @@ public class XssSecurityTest {
             .then()
                 .statusCode(201)
                 .extract()
-                .path("id"));
+                .path("id")).longValue();
 
         given()
                 .contentType(ContentType.JSON)
